+++ examples/openapi/app/source.ts
@@ -3,7 +3,6 @@ import { createMDXSource, defaultSchemas } from 'fumadocs-mdx';
 import { loader } from 'fumadocs-core/source';
 import { attachFile, createOpenAPI } from 'fumadocs-openapi/server';
 import { z } from 'zod';
-import * as path from 'node:path';
 
 const frontmatter = defaultSchemas.frontmatter.extend({
   method: z.string().optional(),
@@ -22,6 +21,4 @@ export const { getPage, getPages, pageTree } = loader({
   },
 });
 
-export const openapi = createOpenAPI({
-  documentOrPath: path.resolve('./unkey.json'),
-});
+export const openapi = createOpenAPI({});