+++ packages/openapi/src/utils/generate-document.ts
@@ -1,31 +1,55 @@
 import { dump } from 'js-yaml';
+import { type OpenAPIV3 as OpenAPI } from 'openapi-types';
+import Slugger from 'github-slugger';
+import { type TableOfContents } from 'fumadocs-core/server';
+import { type StructuredData } from 'fumadocs-core/mdx-plugins';
+import { type ApiPageProps } from '@/server/api-page';
 import type { DocumentContext, GenerateOptions } from '@/generate';
+import { idToTitle } from '@/utils/id-to-title';
+
+interface StaticData {
+  toc: TableOfContents;
+  structuredData: StructuredData;
+}
 
 export function generateDocument(
-  content: string,
-  options: GenerateOptions,
-  frontmatter: {
+  options: GenerateOptions & {
+    dereferenced: OpenAPI.Document;
+    page: ApiPageProps;
+
     title: string;
     description?: string;
     context: DocumentContext;
   },
 ): string {
+  const { frontmatter } = options;
   const out: string[] = [];
+  const extend = frontmatter?.(
+    options.title,
+    options.description,
+    options.context,
+  );
+
+  let meta: object | undefined;
+  if (options.context.type === 'operation') {
+    meta = {
+      method: options.context.endpoint.method,
+      route: options.context.route.path,
+    };
+  }
+
+  const data = generateStaticData(options.dereferenced, options.page);
+
   const banner = dump({
-    title: frontmatter.title,
-    description: frontmatter.description,
+    title: options.title,
+    description: options.description,
     full: true,
-    ...(frontmatter.context.type === 'operation'
-      ? {
-          method: frontmatter.context.endpoint.method,
-          route: frontmatter.context.route.path,
-        }
-      : undefined),
-    ...options.frontmatter?.(
-      frontmatter.title,
-      frontmatter.description,
-      frontmatter.context,
-    ),
+    ...extend,
+    _openapi: {
+      ...meta,
+      ...data,
+      ...(extend?._openapi as object | undefined),
+    },
   }).trim();
   if (banner.length > 0) out.push(`---\n${banner}\n---`);
 
@@ -40,7 +64,50 @@ export function generateDocument(
     out.push(imports);
   }
 
-  out.push(content);
+  out.push(pageContent(options.page));
 
   return out.join('\n\n');
+}
+
+function generateStaticData(
+  dereferenced: OpenAPI.Document,
+  props: ApiPageProps,
+): StaticData {
+  const slugger = new Slugger();
+  const toc: TableOfContents = [];
+  const structuredData: StructuredData = { headings: [], contents: [] };
+
+  for (const item of props.operations) {
+    const operation = dereferenced.paths[item.path]?.[item.method];
+    if (!operation) continue;
+
+    if (props.hasHead && operation.operationId) {
+      const title =
+        operation.summary ??
+        (operation.operationId ? idToTitle(operation.operationId) : item.path);
+      const id = slugger.slug(title);
+
+      toc.push({
+        depth: 2,
+        title,
+        url: `#${id}`,
+      });
+      structuredData.headings.push({
+        content: title,
+        id,
+      });
+    }
+
+    if (operation.description)
+      structuredData.contents.push({
+        content: operation.description,
+        heading: structuredData.headings.at(-1)?.id,
+      });
+  }
+
+  return { toc, structuredData };
+}
+
+function pageContent(props: ApiPageProps): string {
+  return `<APIPage document={${JSON.stringify(props.document)}} operations={${JSON.stringify(props.operations)}} hasHead={${JSON.stringify(props.hasHead)}} />`;
 }