+++ packages/ui/src/mdx.tsx
@@ -12,11 +12,18 @@ import { Card, Cards } from '@/components/card';
 import { Callout } from '@/components/callout';
 import { Heading } from '@/components/heading';
 import { defaultImageSizes } from '@/utils/shared';
+import { cn } from '@/utils/cn';
 
 declare const { Pre }: typeof import('./mdx.client');
 
 function Image(props: ImgHTMLAttributes<HTMLImageElement>): React.ReactElement {
-  return <NextImage sizes={defaultImageSizes} {...(props as ImageProps)} />;
+  return (
+    <NextImage
+      sizes={defaultImageSizes}
+      {...(props as ImageProps)}
+      className={cn('rounded-lg', props.className)}
+    />
+  );
 }
 
 function Table(