+++ packages/content-collections/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @fumadocs/content-collections
 
+## 1.1.2
+
+### Patch Changes
+
+- 758013f: Use Fumadocs Remark Image instead of `rehype-img-size`
+- Updated dependencies [36b771b]
+- Updated dependencies [61b91fa]
+  - fumadocs-core@13.2.2
+
+## 1.1.1
+
+### Patch Changes
+
+- c7aa090: Improve Fumadocs OpenAPI support
+- Updated dependencies [17fa173]
+  - fumadocs-core@13.2.1
+
 ## 1.1.0
 
 ### Minor Changes